requirements.txt
google-genai>=0.3.0
pyyaml>=6.0
rich>=13.0.0
click>=8.1.0
python-dotenv>=1.0.0
questionary>=2.0.0
